Fixed Review Comments + UI
